muser: fix libmuser_mmap_dev()
--Fix addr passed to vm_insert_page()
The addr passed to vm_insert_page() should start from vm_start and
incremented by PAGE_SIZE
hence 'i' must be used in the calculation and not cur_pgidx(which is the
index to fond the page)

--Add new pages index to new_pgs
new_pgs[] is an array of newly allocated pages, it is used for the
cleanup in case of failure. It must be populated by the page_index of
new allocated pages.

Signed-off-by: Swapnil Ingle <swapnil.ingle@nutanix.com>