experimental support for Python bindings plus sample
This patch introduces experimental support for Python bindings, plus the GPIO
sample in Python.

There are many issues with extending Python (from module stuff to error
handling), the most important one is the inability to modify the buf argument
to the region access callbacks when a region is read from. For now I've resorted
to returning the value since in libmuser we limit accesses to 8 bytes (maybe
we should consider that for the C API as well?). Also, the fact that we don't
pass the region being accessed as an argument to a region access callback
results in clunky code (see the REGION_WRAP macro).

Signed-off-by: Thanos Makatos <thanos.makatos@nutanix.com>