move PCI capabilities inside lm_pci_info_t
PCI capabilities are specific to PCI so they should be part of the struct
specific to that device. lm_ctx keeps a copy of this struct and this doesn't
make sense as some of the information contained there are not used at all as
they're stored elsewhere (e.g. class code). All such information, including
the now added PCI capabilities, will be cleaned up in a future patch.

Signed-off-by: Thanos Makatos <thanos.makatos@nutanix.com>