Split compilation unit of muser shared library
This creates several smaller libraries out of the compilation units of
the muser userspace library. It then links them together to build the
shared library.

In this process, we also generate unit test version of the libraries
which include --coverage in the compilation and linking of those files.
Those will be statically linked to unit tests later on.

Signed-off-by: Felipe Franciosi <felipe@nutanix.com>