mmap fd instead of grabbing pages when receiving DMA region registration
This enables huge pages to work. Also, it avoids requiring having to pin memory.

The solution is not 100% correct because since we no longer call vfio_pin_pages,
VFIO doesn't send the DMA registration event and we end up getting duplicate
regions which we reject (this is because of our internal DMA bookkeeping
implementation). We need to work with the community to find a solution.

fixes #28
fixes #29
fixes #38

Signed-off-by: Thanos Makatos <thanos.makatos@nutanix.com>