add support PCI vendor-specific capability
The PCI vendor-specific capability is blindly read/written by the
library. It is possible that the user might want to intercept accesses
to it, in which case we'll have to add callback. The best way to do this
to introduce a new function that configures callbacks for the PCI
capabilities, e.g.

    typedef ssize_t (vfu_cap_access_t) (void *pvt, uint8_t id,
        char *buf, size_t count, loff_t offset, bool is_write);

    vfu_pci_cap_set_cb(vfu_ctx-T *vfu_ctx,  uint8_t cap_id,
	vfu_cap_access_t *cb);

This way the existing API won't have to change.

Signed-off-by: Thanos Makatos <thanos.makatos@nutanix.com>