return region capabilities a la VFIO
This patch returns region capabilities the same way VFIO does: if argsz
is not large enough then it returns only region info and sets argsz to
what it should be in order to fit the capabilities, the client then
retries with a large enough argsz. The protocol specification has been
updated as well.

Plus unit tests.

Signed-off-by: Thanos Makatos <thanos.makatos@nutanix.com>