remove vfu_get_pci_non_std_config_space()
Remove this API as well as vfu_pci_non_std_config_space_t. It's at best
confusing to try to represent this area as if it's not just a normal part of the
overall config space, and we don't really want an additional API for the
extended space past either.

Signed-off-by: John Levon <john.levon@nutanix.com>