add vfu_reset_ctx()
When we fail to send or recv from the socket due to the client disappearing in
some manner, call into vfu_reset_ctx() to clean up the connection fd, allowing a
subsequent vfu_attach_ctx() to work.

If we get 0 bytes from recv[msg](), this is reported by the transport as ENOMSG,
and is a normal EOF condition.

We can also get ECONNRESET: this can happen when we've written unacknowledged
data to the socket, the client side socket is closed, and we try a subsequent
read.

Finally, we can get a short read or write. Our handling of these still has
issues, but for now we'll presume this means the client has gone too.

Signed-off-by: John Levon <john.levon@nutanix.com>