re-work SG dirty tracking
Move SG dirtying to vfu_unmap_sg(): as we no longer track SGs ourselves, doing
this in vfu_map_sg() is no longer the right place.

Note that the lack of tracking implies that any SGs must be unmapped before the
final stop and copy phase. To avoid the need for this, add vfu_mark_sg_dirty():
this allows a consumer to mark a region as dirty explicitly without needing to
unmap it. Currently it's the same as vfu_unmap_sg(), but that's an
implementation detail.

Signed-off-by: John Levon <john.levon@nutanix.com>