samples/client.c: fix implicit casting of const char pointers
samples/client.c implicitly casts const char * to char * in a couple of
places - as such discards the const qualifier. QEMU complains about this
as it builds with -Werror=discarded-qualifiers

This patch declares irq_to_str as an array of const char pointers. It also
casts a "migrate_to() -> _argv" member explicitly

Signed-off-by: Jagannathan Raman <jag.raman@oracle.com>