update to latest version of the spec This pulls in the version of the spec that was submitted to qemu-devel: https://lists.nongnu.org/archive/html/qemu-devel/2023-06/msg06567.html
This version of the spec removes live migration support as it will be
added later on following v2 of the VFIO migration protocol. However,
we will not remove the code for live migration v1 until we implement v2.

We do not implement all of the other new changes in this spec update but
are otherwise compatible.

Signed-off-by: William Henderson <william.henderson@nutanix.com>