samples: improve gpio-pci-idio-16 emulation
Newer kernels implement read register caching for the gpio-pci-idio-16 which
means that the pin counter is never incremented, making it impossible to change
the input register for testing.

Improve the gpio-pci-idio-16 emulation so that the input and output registers
match those of the real hardware, and update the logic so that toggling an
output line 3 times will toggle the corresponding input line to allow testing
individual inputs and outputs.

Signed-off-by: Mark Cave-Ayland <mark.caveayland@nutanix.com>